UXD-805 docs(ListBox): refactor stories and tests
