+++ packages/ListBox/stories/ListBox.stories.mdx
@@ -1,7 +1,7 @@
 import { Meta } from "@storybook/addon-docs/blocks";
 import { getStoryName } from "storybook/storyTree";
 import MDX from "./ListBox.mdx";
-import ListBox from "../src/";
+import ListBox from "../src";
 
 export const params = {
   options: {
