+++ packages/ListBox/stories/examples/OnChange.js
@@ -1,20 +1,26 @@
 import React from "react";
 import RJT from "react-json-tree";
-import ListBox from "../../../src";
+import StoryHeading from "storybook/components/StoryHeading";
+import { Rule } from "storybook/assets/styles/common.styles";
+import ListBox from "../../src";
 
 const propTypes = {};
 const defaultProps = {};
 
-const ParametersExample = ({ state, signature }) => {
+const ParametersExample = ({ state, signature, isSingle = false }) => {
   return (
     <div style={{ padding: "8px", border: "1px solid #CCC", "margin-bottom": "8px", "border-radius": "3px" }}>
-      <span style={{ "font-size": "14px", color: " #333" }}>Single select onChange signature:</span>
+      <span style={{ "font-size": "14px", color: " #333" }}>
+        {isSingle ? "Single select" : "Multi select"} onChange signature:
+      </span>
       <hr />
       <code
         style={{
+          display: "block",
           "font-size": "13px",
           color: "#785CBA",
           padding: "8px",
+          "margin-bottom": "8px",
           "font-weight": "600",
           "border-radius": "3px",
           background: "#EEE",
@@ -23,8 +29,6 @@ const ParametersExample = ({ state, signature }) => {
       >
         {signature}
       </code>
-      <br />
-      <br />
       {state ? <RJT hideRoot data={state} /> : null}
     </div>
   );
@@ -57,18 +61,17 @@ export default function OnChange() {
 
   return (
     <React.Fragment>
-      <div style={{ padding: "16px" }}>
-        <h2>Single selection</h2>
-        <ParametersExample state={singleState} signature={signatureSingleStr} />
+      <div>
+        <StoryHeading level={2}>Single selection</StoryHeading>
+        <ParametersExample state={singleState} signature={signatureSingleStr} isSingle />
         <ListBox onChange={handleSingleChange}>
           <ListBox.Option>One</ListBox.Option>
           <ListBox.Option>Two</ListBox.Option>
           <ListBox.Option>Three</ListBox.Option>
         </ListBox>
         <br />
-        <br />
-        <h2>Single selection with footer</h2>
-        <ParametersExample state={singleStateWithFooter} signature={signatureSingleStr} />
+        <StoryHeading level={2}>Single selection with footer</StoryHeading>
+        <ParametersExample state={singleStateWithFooter} signature={signatureSingleStr} isSingle />
         <ListBox>
           <ListBox.Option>One</ListBox.Option>
           <ListBox.Option>Two</ListBox.Option>
@@ -76,9 +79,9 @@ export default function OnChange() {
           <ListBox.Footer onClickAccept={handleSingleStateWithFooter} />
         </ListBox>
       </div>
-      <hr />
-      <div style={{ padding: "16px" }}>
-        <h2>Multiple selection with footer</h2>
+      <Rule />
+      <div>
+        <StoryHeading level={2}>Multiple selection with footer</StoryHeading>
         <ParametersExample state={multipleState} signature={signatureMultipleStr} />
         <ListBox isMulti onChange={handleMultipleChange}>
           <ListBox.Option>One</ListBox.Option>
@@ -89,8 +92,7 @@ export default function OnChange() {
           <ListBox.Option>Six</ListBox.Option>
         </ListBox>
         <br />
-        <br />
-        <h2>Multiple selection with footer</h2>
+        <StoryHeading level={2}>Multiple selection with footer</StoryHeading>
         <ParametersExample state={multiStateWithFooter} signature={signatureMultipleStr} />
         <ListBox isMulti>
           <ListBox.Option>One</ListBox.Option>
